backend/requirements.txt
Django==3.1.1
django-cors-headers
djangorestframework-simplejwt
djangorestframework==3.11.1
django-filter==2.3.0
celery==4.4.7
django-celery-beat==2.0
django-celery-results==1.2.1
scrapy
psycopg2==2.8.6
requests